add addison anylyzer

+++ Parser/DataAccess/Models/AdvertCarPrice.cs
@@ -5,7 +5,7 @@ namespace DataAccess.Models
     public class AdvertCarPrice
     {
         public int Id { get; set; }
-        public int Value { get; set; }
+        public double Value { get; set; }
         public DateTime DateTime { get; set; }
 
         public int AdvertCarId { get; set; }
